// apps/seeker-portal/__generated__/DashboardHeaderQuery.graphql.ts
export type DashboardHeaderQuery$variables = Record<PropertyKey, never>;
export type DashboardHeaderQuery$data = {
  readonly viewer: {
    readonly __typename: "Account";
    readonly " $fragmentSpreads": FragmentRefs<"AuthNavigationFragment" | "IncompleteProfileBannerFragment">;
  } | {
    readonly __typename: "NotAuthenticatedError";
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
};
export type DashboardHeaderQuery = {
  response: DashboardHeaderQuery$data;
  variables: DashboardHeaderQuery$variables;
};